server/ml_service/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.18
pydantic>=1.10
python-multipart>=0.0.5
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24
openai>=0.27.0
pytest>=7.0
requests>=2.28
httpx>=0.24